Usd interfaces for the clients
