+++ internal/clients/idenfy/interface.go
@@ -1,5 +1,11 @@
 package idenfy
 
+import (
+	"context"
+
+	"example.com/tfgrid-kyc-service/internal/models"
+)
+
 type IdenfyConfig interface {
 	GetBaseURL() string
 	GetCallbackUrl() string
@@ -9,4 +15,9 @@ type IdenfyConfig interface {
 	GetAPIKey() string
 	GetAPISecret() string
 	GetCallbackSignKey() string
+}
+
+type IdenfyClient interface {
+	CreateVerificationSession(ctx context.Context, clientID string) (models.Token, error)
+	VerifyCallbackSignature(ctx context.Context, body []byte, sigHeader string) error
 }
