+++ internal/repository/interface.go
